% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flsa_dp}
\alias{flsa_dp}
\title{Univariate total variation denoising}
\usage{
flsa_dp(y, lambda)
}
\arguments{
\item{y}{Numeric vector of observations.}

\item{lambda}{Regularization parameter.}
}
\value{
Numeric vector of the same length as y, containing denoised
	observations.
}
\description{
Total variation denoising is performed at a fixed tuning parameter value
using the linear-time dynamic programming approach of Johnson (2013).
}
\keyword{internal}
