% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flsa_dp_weighted}
\alias{flsa_dp_weighted}
\title{Weighted univariate total variation denoising}
\usage{
flsa_dp_weighted(y, lambda, weights)
}
\arguments{
\item{y}{Numeric vector of observations.}

\item{lambda}{Regularization parameter.}

\item{weights}{Numeric vector of the same length as y, containing
observation weights.}
}
\value{
Numeric vector of the same length as y, containing denoised
	observations.
}
\description{
Total variation denoising is performed at a fixed tuning parameter
value using the linear-time dynamic programming approach of Johnson
(2013), subject to observation weights
}
\keyword{internal}
