% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvdenoising-package.R
\docType{package}
\name{tvdenoising-package}
\alias{tvdenoising-package}
\title{tvdenoising: Univariate Total Variation Denoising}
\description{
Total variation denoising can be used to approximate a given sequence of noisy observations by a piecewise constant sequence, with adaptively-chosen break points. An efficient linear-time algorithm for total variation denoising is provided here, based on Johnson (2013) \doi{10.1080/10618600.2012.681238}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/glmgen/tvdenoising}
  \item \url{https://glmgen.github.io/tvdenoising/}
  \item Report bugs at \url{https://github.com/glmgen/tvdenoising/issues}
}

}
\author{
\strong{Maintainer}: Ryan Tibshirani \email{ryantibs@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Addison Hu [contributor]
}

}
\keyword{internal}
