% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvdenoising.R
\name{tvdenoising}
\alias{tvdenoising}
\title{Univariate total variation denoising}
\usage{
tvdenoising(y, lambda, weights = NULL)
}
\arguments{
\item{y}{Numeric vector of observations to be denoised.}

\item{lambda}{Regularization parameter.}

\item{weights}{Numeric vector of the same length as y, containing
   observation weights.  The default is `NULL`, which corresponds to unity
weights.}
}
\value{
Denoised observations.
}
\description{
Total variation denoising is performed at a fixed tuning parameter value
using the linear-time dynamic programming approach of Johnson (2013).
Observation weights may be optionally specified.
}
\references{
Johnson (2013), "A dynamic programming algorithm for the fused
 lasso and L0-segmentation."
}
