% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvdenoising-package.R
\docType{package}
\name{tvdenoising-package}
\alias{tvdenoising-package}
\title{tvdenoising: Univariate total variation denoising via dynamic programming}
\description{
Implementation of univariate total variation denoising (aka fused lasso signal approximator) using the linear-time dynamic programming approach of Johnson (2013).
}
\author{
\strong{Maintainer}: Addison Hu \email{mail@huisaddison.com}

Authors:
\itemize{
  \item Ryan Tibshirani
}

}
\keyword{internal}
