% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvdenoising.R
\name{tvdenoising}
\alias{tvdenoising}
\title{Univariate total variation denoising}
\usage{
tvdenoising(y, lambda, weights = NULL)
}
\arguments{
\item{y}{Vector of observations to be denoised.}

\item{lambda}{Regularization parameter value. Must be >= 0.}

\item{weights}{Vector of observation weights. The default is `NULL`, which 
corresponds to unity weights. This vector must have the same length as `y`.}
}
\value{
Vector of denoised observations.
}
\description{
Denoises a sequence of observations by solving the univariate total variation 
denoising optimization problem at a given regularization level.
}
\examples{
y <- c(rnorm(30), rnorm(40, 2), rnorm(30))
yhat <- tvdenoising(y, 10)
plot(y, pch = 16)
lines(yhat, col = 2)
}
\references{
Johnson (2013), "A dynamic programming algorithm for the fused
 lasso and L0-segmentation."
}
